folded load_config.rs into parent module
